+++ lib/cv.js
@@ -34,6 +34,11 @@ exports.register = function (server, option, next) {
             mode: 'try',
             strategy: 'jwt'
           },
+          payload: {
+             output: 'stream',
+             parse: true,
+             allow: 'multipart/form-data'
+         },
           handler: uploadCv
         }
       }
