+++ lib/handlers/cv/upload_cv.js
@@ -1,18 +1,35 @@
 'use strict';
 
 /**
-* Upload the file to Google Drive and update the property cv of the candidate
-* the payload from the form is {id: theIdOfTheCandidate, select-cv: BufferOfTheCv}
+* Upload the file to Google Drive and update the property cvDocumentLink of the candidate
+* the payload from the form is {idCandidate: theIdOfTheCandidate, selectedCv: BufferOfTheCv}
 */
 
-
-const clientES     = require('../../es.js');
-const Google       = require('googleapis');
-const OAuth2       = Google.auth.OAuth2;
-const Oauth2Client = new OAuth2(process.env.GOOGLE_CLIENT_ID, process.env.GOOGLE_CLIENT_SECRET);
+const updloadBuffer = require('./helpers/upload_buffer');
+const fs = require('fs');
 
 module.exports = function (request, reply) {
 
-  console.log(request.payload);
-  return reply('upload cv!')
+  let data = [];
+  let chunk;
+  let stream = request.payload.selectedCv;
+
+  stream.on('readable', () => {
+
+    while ((chunk=stream.read()) != null) {
+
+      data.push(chunk)
+    }
+
+  })
+
+  stream.on('end',function(){
+
+    const binary = Buffer.concat(data);
+    updloadBuffer(binary, stream.hapi.filename, request.payload.idCandidate, function(response) {
+
+     return reply.redirect('/candidate/' + request.payload.idCandidate);
+   })
+  });
+
 }
